Add metaKey to markdown-textarea, for macos
On MacOS, Ctrl is less commonly used than Command (⌘).
Javascript expresses command as `metaKey`.

This allows for _either_ Ctrl or Command to be used in the Markdown text
area.

Note that on Windows, on some browsers, the "windows" key is labeled as meta, so it would
work on windows as well.

http://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/metaKey