Use Join Lemmy hostname as link text in footer (#1737)
* Use Join Lemmy URL as link text in footer when there's a user

I think it makes sense to keep a link to the Join Lemmy website even
though there's a logged in user. There's plenty of other useful info on
that website beyond just joining Lemmy.

Using the URL avoids the phrase "Join Lemmy" and should indicate that
you're navigating to said website.

* Extract hostname from the existing URL constant instead

* Just use hostname of join lemmy site in footer

To maintain consistency whether logged in or not.