Host address environment variable change (#3669)
* Host address environment variable change

Backend and frontend address can be specified as url (without path),
protocol is optional and the default depends on LEMMY_UI_HTTPS. This
allows to use different protocols for the frontend and backend, and the
internal and external backend addresses.

If the frontend is unspecified, protocol and host values from the
request headers are used instead.

Frontend urls are used for the meta tags set by `HtmlTags` and to detect
local urls. Backend urls are used for rss feeds and pictrs images.

This also removes helper methods that return different hosts based on
`isBrowser()`, SSR and CSR should always render the same urls.

* Run prettier on README.md